+++ templates/express/server.js
@@ -1,10 +1,7 @@
 import { createRequestHandler } from "@remix-run/express";
-import { installGlobals } from "@remix-run/node";
 import compression from "compression";
 import express from "express";
 import morgan from "morgan";
-
-installGlobals();
 
 const viteDevServer =
   process.env.NODE_ENV === "production"