+++ templates/remix-javascript/vite.config.js
@@ -1,9 +1,14 @@
 import { vitePlugin as remix } from "@remix-run/dev";
-import { installGlobals } from "@remix-run/node";
 import { defineConfig } from "vite";
 
-installGlobals();
-
 export default defineConfig({
-  plugins: [remix()],
+  plugins: [
+    remix({
+      future: {
+        v3_fetcherPersist: true,
+        v3_relativeSplatPath: true,
+        v3_throwAbortReason: true,
+      },
+    }),
+  ],
 });