+++ packages/remix-serve/CHANGELOG.md
@@ -1,5 +1,14 @@
 # `@remix-run/serve`
 
+## 2.10.0
+
+### Patch Changes
+
+- Upgrade `express` dependency to `^4.19.2` ([#9184](https://github.com/remix-run/remix/pull/9184))
+- Updated dependencies:
+  - `@remix-run/express@2.10.0`
+  - `@remix-run/node@2.10.0`
+
 ## 2.9.2
 
 ### Patch Changes